more accurate worst-case timekeeping
The old method of maintaining frame rate statistics had an odd behavior
in the case of really long frames: it would almost immediately dispose
of timings that were longer than the sliding average window, which means
you would only see the really bad timing for a single frame. Since these
are the timings you're probably most interested in, this seems
undesirable.

To get around that, this adds a minimum queue size to the timer queue.
The initial size I chose is 100 frames: this ensures that, even in
normal operating conditions, the bad timing will be shown for 100
frames, which should at least be long enough for someone to notice it.

This change also refactors the timing statistics keeper into its own
class, and adds the ability to get best/average/worst case timing for
networking as well as engine frames.

The old API is preserved here by copying data from the new
TimingStatistics class into the places in the Timer classes that other
things go looking for it.