store DiscreteParameter values as strings when relevant
When DiscreteParameters are instantiated with options, the user
of the API is almost certainly trying to use it to represent an
enumeration. DiscreteParameter values are serialized as integer
indexes into that options array which means, if the set of values
is changed, all existing serialized values are invalidated.

This commit changes DiscreteParameter to store values as strings
in the case where it is instantiated with an options array.
Existing files will be correctly loaded, whether or not the
parameter has an options array, and if the file is re-saved
the string option will be stored instead.