don't store information about MIDI surfaces in project files
We were previously only storing the description of the surface, and
we weren't loading anything when we loaded from the project file. This
made for an amusing amplification bug where we would just make a new
surface entry that stuck around forever if you ever saved a project with
an APC40MkII attached to your computer, adding a new one every time you
save.

Since this information wasn't being used before (we were only writing an
immutable property of the hardware itself, so nothing useful, and we
were loading no information at all) this change just removes it from
project files. If, in the future, we want to store and track information
about attached surfaces, we can re-add it, but for now, YAGNI.